separated cli mod in different files
